+++ src/main/java/com/r3m25/client/Application.java
@@ -1,8 +1,8 @@
-package com.r3m25.acl.client;
+package com.r3m25.client;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.feign.EnableFeignClients;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
