+++ src/main/java/com/r3m25/client/client/domain/User.java
@@ -1,4 +1,4 @@
-package com.r3m25.acl.client.domain;
+package com.r3m25.client.client.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,15 +9,13 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Customer {
-
-    private String customerId;
+public class User {
+    private Integer id;
     private String name;
     private String username;
     private String email;
-    private String address;
+    private Address address;
     private String phone;
     private String webSite;
-    private String company;
-
+    private Company company;
 }
