+++ src/main/java/com/r3m25/client/mapper/anotation/CustomerAddress.java
@@ -1,0 +1,14 @@
+package com.r3m25.client.mapper.anotation;
+
+import org.mapstruct.Qualifier;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Qualifier
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.CLASS)
+public @interface CustomerAddress {
+}
