+++ src/main/resources/application.yml
@@ -3,6 +3,7 @@ name: acl.client
 server:
   servlet:
     context-path: /r3m25/api/${name}/v1
+  port: 8090
 
 jsonplaceholder:
   url: https://jsonplaceholder.typicode.com/
